Support in Erlang's SSL library for GRiSP's Secure Element
With this patch there is not need anymore for OpenSSL crypto engines.

Usage:

    Key = #{algorithm => ecdsa, sign_fun => {my_module, my_sign_fun}}.

Then add {key, Key} to the SSL options (server and/or client possible). The
sign_fun has arity one and will get the to be signed Message (TLS 1.3) or
{digest, Digest} (TLS 1.2), analogous to ssl:sign/3. The intended use here
is providing a callback for grisp_cryptoauth:sign/2. While this patch makes
heavy use of the existing infrastructure for crypto engines it is
completely backwards compatible, e.g. crypto engines still work.

Remember to use SSL options for server and client properly:

    * {signature_algs, [{sha256, ecdsa}]}                  %% TLS 1.2
    * {signature_algs_cert, [ecdsa_secp256r1_sha256]}      %% TLS 1.3

The reason is that the Secure Element only supports sha256 and ECDSA with the P-256 curve.

Tested with OTP 23.3.4.5, cowboy 2.9.0 and gun 1.3.3 (latest hex versions in August 2021).